refactor: make compiler name module authoritative
