+++ jazz-next/jazz-next.cabal
@@ -21,7 +21,6 @@ library
       JazzNext.Compiler.Diagnostics
       JazzNext.Compiler.Driver
       JazzNext.Compiler.FractionalLiteral
-      JazzNext.Compiler.Identifier
       JazzNext.Compiler.ModuleGraph
       JazzNext.Compiler.ModuleInterface
       JazzNext.Compiler.ModuleCompiler
