+++ jazz-next/src/JazzNext/Compiler/Parser/Expression.hs
@@ -21,7 +21,7 @@ import JazzNext.Compiler.FractionalLiteral
   ( fractionalLiteralExceedsMagnitude,
     mkFractionalLiteralSource
   )
-import JazzNext.Compiler.Identifier
+import JazzNext.Compiler.Name
   ( identifierText,
     mkIdentifier
   )
