+++ jazz-next/src/JazzNext/Compiler/TypeInference/Solver.hs
@@ -33,8 +33,7 @@ import JazzNext.Compiler.BuiltinCatalog
     numericTypeSupportsRuntimeArithmetic,
     numericTypeSupportsRuntimeComparison
   )
-import JazzNext.Compiler.Identifier (identifierText)
-import JazzNext.Compiler.Name (Name)
+import JazzNext.Compiler.Name (Name, identifierText)
 import JazzNext.Compiler.TypeInference.State
   ( InferState (..),
     SolverState (..),
